+++ odoo/local-src/specific_stock/models/__init__.py
@@ -9,4 +9,3 @@ from . import stock_config_settings
 from . import stock_inventory
 from . import stock_pack_operation
 from . import stock_quant_package
-from . import picking_zone