+++ HISTORY.rst
@@ -9,12 +9,15 @@ Unreleased
 **Data Migration**
 
 * ALCN-1113: get discount_global from AS400 on purchase order lines.
+* TODO update sample and full data on new release
 
 **Features and Improvements**
 
 * ALCN-1100: Set digits precision for product uom to 0
 
 **Bugfixes**
+
+* ALCN-1141: Do the missing load of free products information of supplierinfo
 
 **Build**